Added the WAIVER associated with the UNLICENSE
For the purposes of this explanation, instances of the word "Unlicense" shall
refer to the designers of the document. "unlicense" shall refer to the actual
unlicense document.

Unlicense makes the recommendation that any major contributors to a project
covered by the unlicense should digitally sign a waiver. They even go so far as
to provide a simple template waiver paragraph, which should be included as its
own standalone file.

I have added that particular file.