Fix bug by changing formatting
This PR changes the 4th section’s listing from `(a) […] (b)` and so on to  simply `a) […] b)` because of a bug in some Markdown renderers. Well, a feature, I guess.

`(c)`, even in a list, is converted to the copyright symbol (`&copy;`) specifically on Hugo (which is a static site generator). Now, I don’t know how about you, but I think that converting `(c)` to `©` would create some legal issues. Or, perhaps, it’d just look like garbage. In either way, this is not ideal.